+++ client/src/client/client_connection.rs
@@ -17,21 +17,19 @@
 use std::io;
 
 use futures::Future;
-use tokio_core::reactor::Core;
+use tokio_core::reactor::Handle;
 
 use trust_dns_proto::DnsStreamHandle;
 
 use error::*;
 
 /// Trait for client connections
-pub trait ClientConnection: Sized {
+pub trait ClientConnection: Clone + Send + Sized {
     /// The associated DNS Message stream type.
     type MessageStream;
 
     /// Return the inner Futures items
     ///
     /// Consumes the connection and allows for future based operations afterward.
-    fn unwrap(
-        self,
-    ) -> (Core, Box<Future<Item = Self::MessageStream, Error = io::Error>>, Box<DnsStreamHandle<Error = ClientError>>);
+    fn new_stream(&self, handle: &Handle) -> ClientResult<(Box<Future<Item = Self::MessageStream, Error = io::Error>>, Box<DnsStreamHandle<Error = ClientError>>)>;
 }