+++ client/src/tcp/tcp_client_connection.rs
@@ -20,20 +20,20 @@ use std::time::Duration;
 
 use futures::Future;
 use tokio_core::net::TcpStream;
-use tokio_core::reactor::Core;
+use tokio_core::reactor::Handle;
 use trust_dns_proto::DnsStreamHandle;
 
 use error::*;
-use client::{ClientConnection, ClientStreamHandle};
+use client::ClientConnection;
 use tcp::TcpClientStream;
 
 /// Tcp client connection
 ///
 /// Use with `trust_dns::client::Client` impls
+#[derive(Clone)]
 pub struct TcpClientConnection {
-    io_loop: Core,
-    tcp_client_stream: Box<Future<Item = TcpClientStream<TcpStream>, Error = io::Error>>,
-    client_stream_handle: Box<ClientStreamHandle<Error = ClientError>>,
+    name_server: SocketAddr,
+    timeout: Duration,
 }
 
 impl TcpClientConnection {
@@ -60,14 +60,9 @@ impl TcpClientConnection {
     ///
     /// * `name_server` - address of the name server to use for queries
     pub fn with_timeout(name_server: SocketAddr, timeout: Duration) -> ClientResult<Self> {
-        let io_loop = try!(Core::new());
-        let (tcp_client_stream, handle) =
-            TcpClientStream::<TcpStream>::with_timeout(name_server, &io_loop.handle(), timeout);
-
         Ok(TcpClientConnection {
-            io_loop: io_loop,
-            tcp_client_stream: tcp_client_stream,
-            client_stream_handle: handle,
+            name_server,
+            timeout,
         })
     }
 }
@@ -75,13 +70,17 @@ impl TcpClientConnection {
 impl ClientConnection for TcpClientConnection {
     type MessageStream = TcpClientStream<TcpStream>;
 
-    fn unwrap(
-        self,
-    ) -> (Core, Box<Future<Item = Self::MessageStream, Error = io::Error>>, Box<DnsStreamHandle<Error = ClientError>>) {
-        (
-            self.io_loop,
-            self.tcp_client_stream,
-            self.client_stream_handle,
-        )
+    fn new_stream(
+        &self,
+        handle: &Handle,
+    ) -> ClientResult<
+        (Box<Future<Item = Self::MessageStream, Error = io::Error>>,
+         Box<DnsStreamHandle<Error = ClientError>>),
+    > {
+        let (tcp_client_stream, handle) =
+            TcpClientStream::<TcpStream>::with_timeout(self.name_server, handle, self.timeout);
+
+        Ok((tcp_client_stream, handle))
+
     }
 }