+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Many interfaces moved from `client::ClientStreamHandle` to `trust_dns_proto::DnsStreamHandle`
 - `Message::sign` has been renamed and change to the more general method `Message::finalize`
 - Some `io::Error`s have been converted to `trust_dns_proto::ProtoError`
+- `SyncClient` and `SecureSyncClient` are now `Send`
 
 ### Fixed